requirements.txt
MetaTrader5>=5.0.45
numpy>=1.26
pandas>=2.2
pyarrow>=15.0
scipy>=1.11
matplotlib>=3.8
PyYAML>=6.0
jupyter>=1.0
ipykernel>=6.29
tqdm>=4.66
